@@ -10,7 +10,7 @@
   ],
   "type": "contao-module",
   "homepage": "http://now.metamodel.me/",
-  "license": "LGPL-3.0+",
+  "license": "LGPL-3.0-or-later",
   "authors": [
     {
       "name": "Christian Schiffler",
@@ -32,17 +32,17 @@
     "source": "https://github.com/MetaModels/attribute_alias"
   },
   "require": {
-    "php": ">=5.4",
+    "php": "^5.6 || ^7.0",
     "contao/core": "^3.5.5",
-    "contao-community-alliance/composer-plugin": "~2.0",
-    "metamodels/core": "~2.0@alpha",
-    "contao-community-alliance/dc-general": "~2.0@beta",
-    "menatwork/contao-multicolumnwizard": "~3.3"
+    "contao-community-alliance/composer-plugin": "^2.0",
+    "metamodels/core": "^2.0",
+    "contao-community-alliance/dc-general": "^2.0@beta",
+    "menatwork/contao-multicolumnwizard": "^3.3"
   },
   "require-dev": {
-    "cyberspectrum/contao-toolbox": "~0.6",
-    "metamodels/base-unit-tests": "~1.0",
-    "phpcq/all-tasks": "~1.1"
+    "cyberspectrum/contao-toolbox": "^0.6",
+    "metamodels/base-unit-tests": "^1.0.4",
+    "phpcq/all-tasks": "^1.2"
   },
   "autoload": {
     "psr-0": {
