@@ -8,12 +8,19 @@
  *
  * @license http://www.gnu.org/licenses/lgpl-3.0.html LGPL
  *
- * last-updated: 2013-05-13T22:56:21+02:00
+ * last-updated: 2018-02-06T02:11:28+01:00
  */
 
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['0']    = 'Alias-Felder';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['1']    = 'Bitte wählen Sie eines oder mehrere Attribute, um daraus einen Alias zu bilden.';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute']      = 'Attribute';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['0']     = 'Neuerstellung des Alias erzwingen.';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['1']     = 'Auswählen, um eine Neuerstellung des Alias zu erzwingen, wenn sich eines der abhängigen Felder ändert. Beachten Sie, dass bisherige URLs dadurch ungültig werden können.';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['typeOptions']['alias'] = 'Alias';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['0']           = 'Alias-Felder';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['1']           = 'Bitte wählen Sie eines oder mehrere Attribute, um daraus einen Alias zu bilden.';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_postfix']['0']          = 'Alias-Postfix';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_postfix']['1']          = 'Optionale Angabe eines Postfix für die Erstellung des Alias.';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_prefix']['0']           = 'Alias-Präfix';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_prefix']['1']           = 'Optionale Angabe eines Präfix für die Erstellung des Alias.';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute']             = 'Attribute';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['0']            = 'Neuerstellung des Alias erzwingen.';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['1']            = 'Auswählen, um eine Neuerstellung des Alias zu erzwingen, wenn sich eines der abhängigen Felder ändert. Beachten Sie, dass bisherige URLs dadurch ungültig werden können.';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['select_values']['attributes'] = 'Attribute';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['select_values']['meta']       = 'Metafelder';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['typeOptions']['alias']        = 'Alias';
+
