@@ -8,13 +8,14 @@
  *
  * @license http://www.gnu.org/licenses/lgpl-3.0.html LGPL
  *
- * last-updated: 2013-05-13T22:12:15+02:00
+ * last-updated: 2018-02-06T02:11:28+01:00
  */
 
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['0']           = 'Πεδια ψευδωνυμου';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['1']           = 'Παρακαλούμε επιλέξτε ένα ή περισσότερα χαρακτηριστικά για να συνδυάσετε ενα ψευδωνυμο. ';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute']             = ' Χαρακτηριστικά';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['0']            = 'Force αναγέννηση ψευδωνυμου';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['1']            = 'Ελέγξτε αυτό, αν θέλετε το ψευδώνυμο που θα αναγεννηθεί κάθε φορά που κάποιο από τα πεδία που εξαρτώνται αλλάζει. Σημειώστε ότι αυτό θα ακυρώσει τις παλιές διευθύνσεις URL που βασίζονται στα ψευδώνυμο. ';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['select_values']['attributes'] = ' Χαρακτηριστικά';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['typeOptions']['alias']        = 'Ψευδώνυμο ';
 
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['0']    = 'Πεδια ψευδωνυμου';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['1']    = 'Παρακαλούμε επιλέξτε ένα ή περισσότερα χαρακτηριστικά για να συνδυάσετε ενα ψευδωνυμο. ';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute']      = ' Χαρακτηριστικά';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['0']     = 'Force αναγέννηση ψευδωνυμου';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['1']     = 'Ελέγξτε αυτό, αν θέλετε το ψευδώνυμο που θα αναγεννηθεί κάθε φορά που κάποιο από τα πεδία που εξαρτώνται αλλάζει. Σημειώστε ότι αυτό θα ακυρώσει τις παλιές διευθύνσεις URL που βασίζονται στα ψευδώνυμο. ';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['typeOptions']['alias'] = 'Ψευδώνυμο ';
