@@ -8,12 +8,15 @@
  *
  * @license http://www.gnu.org/licenses/lgpl-3.0.html LGPL
  *
- * last-updated: 2013-07-10T10:03:49+02:00
+ * last-updated: 2018-02-06T02:11:28+01:00
  */
 
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['0']    = 'Champs alias';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['1']    = 'Sélectionner un ou plusieurs attributs pour combiner un alias.';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute']      = 'Attributs';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['0']     = 'Forcer la régénération de l\'alias';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['1']     = 'Cocher ceci, si vous voulez que l\'alias soit régénéré à chaque fois que l\'un des champs dépendant est modifié. Notez que ceci invalidera les anciennes url qui sont basées sur l\'alias';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['typeOptions']['alias'] = 'Alias';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['0']           = 'Champs alias';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['1']           = 'Sélectionner un ou plusieurs attributs pour combiner un alias.';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute']             = 'Attributs';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['0']            = 'Forcer la régénération de l\'alias';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['1']            = 'Cocher ceci, si vous voulez que l\'alias soit régénéré à chaque fois que l\'un des champs dépendant est modifié. Notez que ceci invalidera les anciennes url qui sont basées sur l\'alias';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['select_values']['attributes'] = 'Attributs';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['select_values']['meta']       = 'Champs meta';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['typeOptions']['alias']        = 'Alias';
+
