@@ -8,13 +8,14 @@
  *
  * @license http://www.gnu.org/licenses/lgpl-3.0.html LGPL
  *
- * last-updated: 2013-05-13T22:12:15+02:00
+ * last-updated: 2018-02-06T02:11:28+01:00
  */
 
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['0']           = 'Champs dad alias';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['1']           = 'Tscherna in u plirs attributs che duain vegnir cumbinads ad in alias.';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute']             = 'Attributs';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['0']            = 'Sfurzar la re-creaziun da l\'alias';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['1']            = 'Tscherna questa chaschetta, sche ti vuls che l\'alias vegn re-creà mintga giada sch\'in champ dependent vegn midà. Fa stim che quai vegn a render nunvalid urls veglias che sa basan sin ils alias.';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['select_values']['attributes'] = 'Attributs';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['typeOptions']['alias']        = 'Alias';
 
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['0']    = 'Champs dad alias';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['1']    = 'Tscherna in u plirs attributs che duain vegnir cumbinads ad in alias.';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute']      = 'Attributs';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['0']     = 'Sfurzar la re-creaziun da l\'alias';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['1']     = 'Tscherna questa chaschetta, sche ti vuls che l\'alias vegn re-creà mintga giada sch\'in champ dependent vegn midà. Fa stim che quai vegn a render nunvalid urls veglias che sa basan sin ils alias.';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['typeOptions']['alias'] = 'Alias';
