@@ -8,14 +8,15 @@
  *
  * @license http://www.gnu.org/licenses/lgpl-3.0.html LGPL
  *
- * last-updated: 2014-05-20T12:13:36+02:00
+ * last-updated: 2018-02-06T02:11:28+01:00
  */
 
-
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['0']    = 'Псевдоним поля';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['1']    = 'Выберите один или несколько атрибутов для объединения алиаса.';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute']      = 'Атрибуты';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['0']     = 'Принудительно регенерировать алиас';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['1']     = 'Выберите, для регенерации алиаса каждый раз при изменении любого из подчиненных полей. Обратите внимание, что это лишит вас старых URL-адресов основанных на алиасе.';
-$GLOBALS['TL_LANG']['tl_metamodel_attribute']['typeOptions']['alias'] = 'Алиас';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['0']           = 'Псевдоним поля';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields']['1']           = 'Выберите один или несколько атрибутов для объединения алиаса.';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute']             = 'Атрибуты';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['0']            = 'Принудительно регенерировать алиас';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias']['1']            = 'Выберите, для регенерации алиаса каждый раз при изменении любого из подчиненных полей. Обратите внимание, что это лишит вас старых URL-адресов основанных на алиасе.';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['select_values']['attributes'] = 'Атрибуты';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['select_values']['meta']       = 'Метаполя';
+$GLOBALS['TL_LANG']['tl_metamodel_attribute']['typeOptions']['alias']        = 'Алиас';
 
