@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_alias.
  *
- * (c) 2012-2016 The MetaModels team.
+ * (c) 2012-2018 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,23 +13,24 @@
  * @package    MetaModels
  * @subpackage Tests
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2012-2016 The MetaModels team.
+ * @author     Sven Baumann <baumann.sv@gmail.com>
+ * @copyright  2012-2018 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_alias/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-error_reporting(E_ALL);
+\error_reporting(E_ALL);
 
 function includeIfExists($file)
 {
-    return file_exists($file) ? include $file : false;
+    return \file_exists($file) ? include $file : false;
 }
 
 if (
     // Locally installed dependencies.
-    (!$loader = includeIfExists(__DIR__.'/../vendor/autoload.php'))
+    (!$loader = \includeIfExists(__DIR__.'/../vendor/autoload.php'))
     // We are within an composer install.
-    && (!$loader = includeIfExists(__DIR__.'/../../../autoload.php'))) {
+    && (!$loader = \includeIfExists(__DIR__.'/../../../autoload.php'))) {
     echo 'You must set up the project dependencies, run the following commands:'.PHP_EOL.
         'curl -sS https://getcomposer.org/installer | php'.PHP_EOL.
         'php composer.phar install'.PHP_EOL;
